[change] file naming convention to kebab case on src/argument-types/*
