+++ src/tests/helper/assert_set_iargument_testvalue.ts
@@ -1,5 +1,5 @@
 import { IArgument } from '../../arguments/i-argument'
-import { ArgumentType } from '../../argumentTypes/argumentType'
+import { ArgumentType } from '../../argument-types/argument-type'
 
 /**
  * 引数に与えられたArgumentのテスト用値が引数に与えられた型と一致するか確認します。
